% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fx}
\alias{fx}
\alias{fx.eps}
\title{Half-normal pdf and related functions}
\usage{
fx(x,sigma,w)

       fx.eps(eps, x, par, w)
}
\arguments{
\item{x}{distance}

\item{sigma}{half-normal scale (std deviation)}

\item{eps}{std normal deviate}

\item{par}{parameter vector (beta, beta_eps)}

\item{w}{half-width of transect}
}
\description{
Half-normal pdf and related functions
}
\author{
Jeff Laake
}

