% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{hp}
\alias{hp}
\title{Harbor porpoise vessel survey data}
\format{A data frame with 477 observations on the following 5 variables.
\describe{ \item{object}{sequential object #}
           \item{distance}{perpendicular distance to observed harbor porpoise group}
           \item{size}{size of harbor porpoise group}
           \item{detected}{always 1}
           \item{observer}{always 1}}}
\description{
Line transect sampling data for habor porpoise from a vessel survey in the inland Washington waters
}
\details{
The data include perpendicular distances to observed harbor porpoise groups and the size of the group.
It also includes observer (always 1), detected (always 1) and object (sequential number) that are needed
by mrds package.
}
\examples{
dev.new()
par(mfrow=c(1,2))
data(hp)
# hp=read.delim("hp.txt",header=TRUE)
# fit half-normal detection function with group size as a covariate
modmcds=ddf(dsmodel=~cds(key="hn",formula=~size),data=hp,method="ds")
# show summary of results
xx=summary(modmcds)
xx
# fit mixed effect model with eq (9) and use same truncation as mrds analysis
modmixed=fitadmb(hp,formula=~size,w=443.1635)
summary(modmixed)
# plot results and compute abundances in covered area
Nhat.mcds=xx$Nhat
Nhat.se.mcds=xx$Nhat.se
par(mfrow=c(2,1))
plotfit(hp$distance,w=443.1635,dm=model.matrix(~size,hp[,"size",drop=FALSE]),
       par=modmixed$coeff[1:3])
plot(modmcds,new=FALSE,showpoints=FALSE,pl.den=0,main="")
Nhat.mixed=compute_Nhat(par=modmixed$coeff[1:3],x=hp$distance,w=443.1635,
                dm=model.matrix(~size,hp[,"size",drop=FALSE]))
Nhat.se.mixed=compute_Nhat.se(par=modmixed$coeff[1:3],vcov=modmixed$vcov[1:3,1:3],
                x=hp$distance,w=443.1635,dm=model.matrix(~size,hp[,"size",drop=FALSE]))
formatC(xx$aic-(-2*modmixed$loglik+2*modmixed$npar),digits=2)
paste(formatC(Nhat.mcds,digits=4)," (se = ",formatC(Nhat.se.mcds,digits=2),")",sep="")
paste(formatC(Nhat.mixed,digits=4)," (se = ",formatC(Nhat.se.mixed,digits=2),")",sep="")
}
\keyword{datasets}

