% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RandomScale-package}
\alias{RandomScale}
\alias{RandomScale-package}
\title{Random Scale Detection Functions for Line Transect Data}
\description{
The RandomScale package provides R and AD Model Builder(\url{http://admb-project.org})code to fit random effect and mixed effect models for the scale
of a half-normal detection function to line transect sampling data.
}
\details{
The ADMB TPL files are contained in the Inst directory(\url{https://github.com/jlaake/RandomScale/tree/master/RandomScale/inst}).
Example AMDB DAT files are contained in the Data directory(\url{https://github.com/jlaake/RandomScale/tree/master/RandomScale/Data}).
Version 11 of ADMB is required for the TPL files as distributed but you can modify them by replacing instances of PI with the
value 3.141592654. Executable versions are also provided so the user does not need ADMB but the R2admb package is required because
it is used to extract results from the output file.

If you wish to to install ADMB, you need to install a C++ compiler as well as ADMB. We suggest that
you install gcc. To use the built-in links in the function prepare_admb, admb should be installed to c:\\admb and the gcc
compiler to c:\\mingw. To use different locations you'll need to change function prepare_admb. Unless you use existing
exe's in the package, you'll need to run prepare_admb for each R session.

For Windows install ADMB-11 (\url{http://admb-project.googlecode.com/files/admb-11-linux-gcc4.6-32bit.zip}) and
gcc(\url{http://www.admb-project.org/tools/gcc/gcc452-win32.zip/at_download/file})

For other operating systems see (\url{http://www.admb-project.org/downloads}) and
gcc(\url{http://www.admb-project.org/tools/gcc/}). Note that prepare_admb() only works for Windows.

For more information see \url{https://github.com/downloads/jlaake/ADMB-Examples/distance_random_effect.pdf}. Also the file likelihoods.pdf
in the RandomScale package directory contains a description of the alternative likelihoods including an incorrect likelihood f1 which was
removed from the paper but remains in the code.
}
\examples{
# the following runs the example in Oedekoven et al;
# note that you need to change nreps to 100 in simulation to get the same example
example(fitadmb)
example(hp)
example(simulation)
}
\author{
Jeff Laake
}
\references{
Fournier, D.A., H.J. Skaug, J. Ancheta, J. Ianelli, A. Magnusson, M.N. Maunder, A. Nielsen, and J. Sibert. 2012. AD Model Builder: using automatic differentiation for statistical inference of highly parameterized complex nonlinear models. Optim. Methods Softw. 27:233-249.
}

