\name{fx}
\alias{fx}
\alias{fx.eps}
\title{Half-normal pdf and related functions}
\usage{
  fx(x,sigma,w)

  fx.eps(eps, x, par, w)
}
\arguments{
  \item{x}{distance}

  \item{sigma}{half-normal scale (std deviation)}

  \item{eps}{std normal deviate}

  \item{par}{parameter vector (beta, beta_eps)}

  \item{w}{half-width of transect}

  \item{weps}{range (-weps,weps) for std normal
  integration}
}
\description{
  Half-normal pdf and related functions
}
\author{
  Jeff Laake# pdf - for incorrect likelihood
}

