\name{mu}
\alias{avg_mu}
\alias{mu}
\title{Integration of Half-normal detection}
\usage{
  mu(eps,w,par)

  avg_mu(w,par,weps=5)
}
\arguments{
  \item{x}{distance}

  \item{sigma}{half-normal scale (std deviation)}

  \item{eps}{std normal deviate}

  \item{par}{parameter vector (beta, beta_eps)}

  \item{weps}{range (-weps,weps) for std normal
  integration}
}
\description{
  Integration of Half-normal detection
}
\author{
  Jeff Laake
}

